+++ django_echarts/entities/__init__.py
@@ -3,8 +3,8 @@ from .articles import (
 )
 from .base import DwString
 from .chart_widgets import (NamedCharts, WidgetGetterMixin)
-from .containers import WidgetCollection
 from .html_widgets import (
     LinkItem, Menu, Jumbotron, Nav, Copyright, Message, ValuesPanel, ValueItem, table_css
 )
-from .rows import RowContainer
+from .pages import WidgetCollection
+from .containers import RowContainer, Container